Refactor: Rename director→illustrator, illustrator→renderer
Better reflects what each agent does:
- Illustrator: Creates visual story/storyboard from manuscript
- Renderer: Generates actual images from story

Updated files, types, exports, and documentation.
