+++ ARCHITECTURE.md
@@ -40,8 +40,8 @@ bookbug/
 │   │   │   ├── interpreter.ts    # Extracts StoryBrief from user input
 │   │   │   ├── conversation.ts   # Generates follow-up questions for chat
 │   │   │   ├── author.ts         # Manuscript generation from brief
-│   │   │   ├── director.ts       # Visual story direction from manuscript
-│   │   │   └── illustrator.ts    # Image generation from story
+│   │   │   ├── illustrator.ts    # Visual story direction from manuscript
+│   │   │   └── renderer.ts       # Image generation from story
 │   │   └── pipeline.ts           # Pipeline orchestration
 │   └── cli/
 │       ├── index.ts              # CLI entry point (Commander.js)
@@ -142,7 +142,7 @@ Generates the complete `Story.json` containing manuscript text and visual direct
 
 **Agents:**
 1. **AuthorAgent**: `StoryBlurb` → `Manuscript` (per-page text, summaries, image concepts)
-2. **DirectorAgent**: `Manuscript` → `Story` (visual style guide, shot compositions, beat breakdowns)
+2. **IllustratorAgent**: `Manuscript` → `Story` (visual style guide, shot compositions, beat breakdowns)
 3. **TypographerAgent** *(planned)*: `Story` → `Story` with text styling (font, position, size, decoration per page)
 
 **Output:** `Story.json` (normalized blob with characters lookup, manuscript pages, visual beats)
@@ -224,8 +224,8 @@ export async function executePipeline(
   options: PipelineOptions = {}
 ): Promise<PipelineResult> {
   const manuscript = await authorAgent(brief);
-  const story = await directorAgent(manuscript);
-  const book = await illustratorAgent(story);
+  const story = await illustratorAgent(manuscript);
+  const book = await rendererAgent(story);
   return { brief, manuscript, story, book };
 }
 ```
@@ -343,11 +343,11 @@ inngest.createFunction(
     const manuscript = await step.run("write-manuscript", () =>
       authorAgent(brief)
     );
-    const story = await step.run("direct-story", () =>
-      directorAgent(manuscript)
+    const story = await step.run("illustrate", () =>
+      illustratorAgent(manuscript)
     );
-    const book = await step.run("illustrate", () =>
-      illustratorAgent(story)
+    const book = await step.run("render", () =>
+      rendererAgent(story)
     );
     return book;
   }
